requirements.txt
django==3.1.1
djangorestframework==3.11.1
django_extensions==3.0.9
django-cors-headers==3.5.0
psycopg2==2.8.6
requests==2.24.0
requests_futures==1.0.0
